% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{readHAR}
\alias{readHAR}
\title{Read HAR objects}
\usage{
readHAR(har, ...)
}
\arguments{
\item{har}{a string, list/URL or file containing JSON HAR data}

\item{...}{addtional arguments}
}
\value{
returns a parsed HAR object
}
\description{
Read HAR objects
}
\examples{
har <- readHAR(system.file(package = "HARtools",
                        "exdata", "google.com.har"))
}

